Adicionado Policies, cache e Cors

+++ app/Http/Controllers/ProductsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Product;
 use App\Http\Requests\ProductsRequest as Request;
+use Illuminate\Support\Facades\Auth;
 
 class ProductsController extends Controller
 {
@@ -16,12 +17,19 @@ class ProductsController extends Controller
 
     public function index()
     {
-        return $this->item::all();
+        $minutes = \Carbon\Carbon::now()->addMinutes(1);
+        $products = \Cache::remember('api::product', $minutes, function () {
+            return $this->item::all();
+        });
+        return $products;
     }
 
     public function store(Request $request)
     {
-        return $this->item::create($request->all());
+        \Cache::forget('api::product');
+        $dataForm = $request->all();
+        $dataForm['user_id'] = Auth::user()->id;
+        return $this->item::create($dataForm);
     }
 
     public function show($id)
@@ -31,12 +39,15 @@ class ProductsController extends Controller
 
     public function update(Request $request)
     {
+            \Cache::forget('api::product');
             $this->item->update($request->all());
             return $this->item;
     }
 
     public function destroy($id)
     {
+        $this->authorize('delete', $this->item->find($id));
+        \Cache::forget('api::product');
         $this->item->delete();
         return $this->item;
     }
